2006-08-26  Stefan Jahn  <stefan@lkcc.org>
        * qucsdigi.bat: Closing stdin when running the simulator.

        * qtgeneric_tr.ts, qucs_tr.ts: Updated Turkish translations.
        Thanks to Ozgur.

        * Makefile.am (bin_SCRIPTS): Care about the host system when
        installing the qucsdigi script.

2006-08-26  Stefan Jahn  <stefan@lkcc.org>

        * qucshelp.cpp (setupActions): Made some more user visible
        strings translatable.  Renamed the "Sidebar" into "Table of
        Contents".


git-svn-id: https://qucs.svn.sourceforge.net/svnroot/qucs/trunk@961 b5b04e8c-4942-46c9-ab4f-83783d557d1c

+++ qucs-help/htmldatafetcher.cpp
@@ -273,7 +273,7 @@ QChar HtmlDataFetcher::unicodeFor(QString placeHolder)
   placeHolder.remove(' ');
   placeHolder.remove('\t');
   Q_ASSERT(placeHolder[0] == '&');
-  Q_ASSERT(placeHolder[placeHolder.length() - 1] == ';');
+  Q_ASSERT(placeHolder[(int)placeHolder.length() - 1] == ';');
   placeHolder.remove(placeHolder.length() - 1,1);//remove ';'
   placeHolder.remove(0,1);// remove '&'
   if(placeHolder[0] == '#')
